Updating seed file because of migration problems, changing a blank value on a resource to its actual value.

+++ models/recipe.js
@@ -15,7 +15,7 @@ module.exports = (sequelize, DataTypes) => {
     fat: DataTypes.FLOAT,
     carbs: DataTypes.FLOAT,
     protein: DataTypes.FLOAT,
-    yield: DataTypes.FLOAT
+    yield: DataTypes.INTEGER
   }, {});
   Recipe.associate = function(models) {
     // associations can be defined here
